refactor: Lotto, LottoFactory 를 domain/lotto/ 폴더로 이동
